feat: create new main.ts to run the app in aws elastic beanstalk
