+++ src/main.ts
@@ -1,16 +1,14 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ConfigService } from '@nestjs/config';
-import { Callback, Context, Handler } from 'aws-lambda';
 import helmet from 'helmet';
-import serverlessExpress from '@codegenie/serverless-express';
 
-let server: Handler;
+const PORT = 4000;
 
-async function bootstrap() {
+const bootstrap = async () => {
   const app = await NestFactory.create(AppModule);
   const cfgService = app.get(ConfigService);
-  const env = (name: string, def?: string) => cfgService.get<string>(name);
+  const env = (name: string, def?: string) => cfgService.get<string>(name, def);
 
   app.enableCors({
     origin: env('FRONTEND_ORIGIN'),
@@ -21,18 +19,9 @@ async function bootstrap() {
   });
 
   app.use(helmet());
-  await app.init();
-  const expressApp = app.getHttpAdapter().getInstance();
-  return serverlessExpress({ app: expressApp });
-}
-
-export const handler: Handler = async (
-  event: any,
-  context: Context,
-  callback: Callback,
-) => {
-  console.log('Handler event', event);
-  server = server ?? (await bootstrap());
-  console.log('Server has been started...');
-  return server(event, context, callback);
+  await app.listen(PORT);
 };
+
+bootstrap().then(() => {
+  console.log('The app has been started on port:', PORT);
+});
